Bug fixes for simplify. Don't like results, but it may be working. It's at least not absurd.

+++ orogram/xentropy.pxd
@@ -5,4 +5,4 @@
 # cython: wraparound=False
 # cython: nonecheck=False
 
-cdef float section_crossentropy(float p0, float p1, float q0, float q1, float log_q0, float log_q1) nogil
+cdef float section_crossentropy(float p0, float p1, float q0, float q1, double log_q0, double log_q1) nogil
